backend/requirements.txt
fastapi
uvicorn
pydantic
pillow
numpy
python-multipart
passlib[bcrypt]